// generated/models/BaseResponse_Page_QuestionSetVO_.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Page_QuestionSetVO_ } from './Page_QuestionSetVO_';
export type BaseResponse_Page_QuestionSetVO_ = {
    code?: number;
    data?: Page_QuestionSetVO_;
    message?: string;
};